// docs/contracts/frontend-types.ts
export type VariantCreate = { "name": string; "portionDescription": string; "description": string; "isAvailable": boolean };
export type VariantUpdate = { "expectedVersion": number; "name"?: string; "portionDescription"?: string; "description"?: string; "isAvailable"?: boolean };
export type Variant = { "id": string; "name": string; "portionDescription": string; "description": string; "isAvailable": boolean; "version": number; "deletedAt": (string | null) };
export type DishCreate = { "name": string; "description": string; "imageFileId": (string | null); "isAvailable": boolean; "variants": Array<VariantCreate> };
export type DishUpdate = { "expectedVersion": number; "name"?: string; "description"?: string; "imageFileId"?: (string | null); "isAvailable"?: boolean };